Flux Units Refactor
